# docs/swagger.yaml
definitions:
...
  internal_controllers.AssignRoleRequest:
    properties:
      role_id:
        type: string
    required:
    - role_id
    type: object
...
  internal_controllers.CreateRoleRequest:
    properties:
      description:
        type: string
      display_name:
        maxLength: 100
        minLength: 3
        type: string
      name:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - display_name
    - name
    type: object
...
  internal_controllers.UpdateRoleRequest:
    properties:
      description:
        type: string
      display_name:
        maxLength: 100
        minLength: 3
        type: string
      is_active:
        type: boolean
    required:
    - display_name
    type: object
...
  user-management_internal_utils.Response:
    properties:
      code:
        type: string
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /api/v1/roles:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all roles
      tags:
      - Roles
    post:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Role data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_controllers.CreateRoleRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Create role
      tags:
      - Roles
  /api/v1/roles/{roleId}:
    delete:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete role
      tags:
      - Roles
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get role by ID
      tags:
      - Roles
    put:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: string
      - description: Role data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_controllers.UpdateRoleRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Update role
      tags:
      - Roles
...
  /api/v1/users/{userId}/roles:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get user's roles
      tags:
      - Users
    post:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Role ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_controllers.AssignRoleRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Assign role to user
      tags:
      - Users
  /api/v1/users/{userId}/roles/{roleId}:
    delete:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Remove role from user
      tags:
      - Users
swagger: "2.0"